[DASH] Revert test_fnic name change (#23276)
This PR reverts the name change from #22288 - the tunnel-based FNIC test case is moved back to the file test_fnic.py and the test case added in #22288 is moved to a new file called test_fnic_basic.py
