+++ tests/dash/test_fnic_basic.py
@@ -1,0 +1,115 @@
+import logging
+
+import configs.privatelink_config as pl
+import ptf.testutils as testutils
+import pytest
+from constants import LOCAL_PTF_INTF, REMOTE_PTF_RECV_INTF, REMOTE_PTF_SEND_INTF
+from gnmi_utils import apply_messages
+from packets import rand_udp_port_packets
+from tests.common import config_reload
+
+logger = logging.getLogger(__name__)
+
+pytestmark = [
+    pytest.mark.topology("smartswitch"),
+    pytest.mark.skip_check_dut_health
+]
+
+
+"""
+Test prerequisites:
+- Assign IPs to DPU-NPU dataplane interfaces
+
+Note: It's also necessary for the DPU to learn the neighbor info of the dataplane port to the NPU before any
+DASH configs are programmed. This should be handled automatically by fixture ordering and does not require
+manual steps.
+
+The neighbor info is learned when appling the default route as orchagent will attempt to resolve the next hop IP.
+"""
+
+
+@pytest.fixture(autouse=True)
+def common_setup_teardown(
+    localhost,
+    duthost,
+    ptfhost,
+    dpu_index,
+    skip_config,
+    dpuhosts,
+    set_vxlan_udp_sport_range,
+    # manually invoke setup_npu_dpu to ensure routes are added before DASH configs are programmed
+    setup_npu_dpu,  # noqa: F811
+):
+    if skip_config:
+        yield
+        return
+    dpuhost = dpuhosts[dpu_index]
+    logger.info(pl.ROUTING_TYPE_PL_CONFIG)
+
+    base_config_messages = {
+        **pl.APPLIANCE_FNIC_CONFIG,
+        **pl.ROUTING_TYPE_PL_CONFIG,
+        **pl.ROUTING_TYPE_VNET_CONFIG,
+        **pl.VNET_CONFIG,
+        **pl.ROUTE_GROUP1_CONFIG,
+        **pl.METER_POLICY_V4_CONFIG,
+    }
+    logger.info(base_config_messages)
+
+    apply_messages(localhost, duthost, ptfhost, base_config_messages, dpuhost.dpu_index)
+
+    route_and_mapping_messages = {
+        **pl.PE_VNET_MAPPING_CONFIG,
+        **pl.PE_SUBNET_ROUTE_CONFIG,
+        **pl.VM_VNET_MAPPING_CONFIG,
+        **pl.VM_SUBNET_ROUTE_CONFIG
+    }
+    logger.info(route_and_mapping_messages)
+    apply_messages(localhost, duthost, ptfhost, route_and_mapping_messages, dpuhost.dpu_index)
+
+    # inbound routing not implemented in Pensando SAI yet, so skip route rule programming
+    if 'pensando' not in dpuhost.facts['asic_type']:
+        route_rule_messages = {
+            **pl.VM_VNI_ROUTE_RULE_CONFIG,
+            **pl.INBOUND_VNI_ROUTE_RULE_CONFIG,
+            **pl.TRUSTED_VNI_ROUTE_RULE_CONFIG
+        }
+        logger.info(route_rule_messages)
+        apply_messages(localhost, duthost, ptfhost, route_rule_messages, dpuhost.dpu_index)
+
+    logger.info(pl.ENI_FNIC_CONFIG)
+    apply_messages(localhost, duthost, ptfhost, pl.ENI_FNIC_PL_CONFIG, dpuhost.dpu_index)
+
+    logger.info(pl.ENI_ROUTE_GROUP1_CONFIG)
+    apply_messages(localhost, duthost, ptfhost, pl.ENI_ROUTE_GROUP1_CONFIG, dpuhost.dpu_index)
+
+    yield
+
+    # Route rule removal is broken so config reload to cleanup for now
+    # https://github.com/sonic-net/sonic-buildimage/issues/23590
+    if 'pensando' in dpuhost.facts['asic_type']:
+        apply_messages(localhost, duthost, ptfhost, pl.ENI_ROUTE_GROUP1_CONFIG, dpuhost.dpu_index, False)
+        apply_messages(localhost, duthost, ptfhost, pl.ENI_FNIC_CONFIG, dpuhost.dpu_index, False)
+        apply_messages(localhost, duthost, ptfhost, route_and_mapping_messages, dpuhost.dpu_index, False)
+        apply_messages(localhost, duthost, ptfhost, base_config_messages, dpuhost.dpu_index, False)
+    else:
+        config_reload(dpuhost, safe_reload=True, yang_validate=False)
+
+
+@pytest.mark.parametrize("encap_proto", ["vxlan", "gre"])
+def test_fnic(ptfadapter, dash_pl_config, encap_proto):
+    num_packets = 5
+    pkt_sets = list()
+
+    for _ in range(num_packets):
+        vm_to_dpu_pkt, exp_dpu_to_pe_pkt, pe_to_dpu_pkt, exp_dpu_to_vm_pkt = rand_udp_port_packets(
+            dash_pl_config, floating_nic=True, outbound_vni=pl.VNET1_VNI, outbound_encap=encap_proto
+        )
+        pkt_sets.append((vm_to_dpu_pkt, exp_dpu_to_pe_pkt, pe_to_dpu_pkt, exp_dpu_to_vm_pkt))
+
+    ptfadapter.dataplane.flush()
+    for vm_to_dpu_pkt, exp_dpu_to_pe_pkt, pe_to_dpu_pkt, exp_dpu_to_vm_pkt in pkt_sets:
+        testutils.send(ptfadapter, dash_pl_config[LOCAL_PTF_INTF], vm_to_dpu_pkt, 1)
+        testutils.verify_packet_any_port(ptfadapter, exp_dpu_to_pe_pkt, dash_pl_config[REMOTE_PTF_RECV_INTF])
+        testutils.send(ptfadapter, dash_pl_config[REMOTE_PTF_SEND_INTF], pe_to_dpu_pkt, 1)
+        testutils.verify_packet(ptfadapter, exp_dpu_to_vm_pkt, dash_pl_config[LOCAL_PTF_INTF])
